Segment item_rando.c and init.c
