+++ base-hack/asm/objects.asm
@@ -1,5 +1,4 @@
 .importobj "obj/src_f3dex2.o"
-.importobj "obj/src_init.o"
 .importobj "obj/src_lib.o"
 .importobj "obj/src_lib_items.o"
 .importobj "obj/src_main.o"
@@ -13,6 +12,17 @@
 .importobj "obj/src_fixes_quality_fixes.o"
 .importobj "obj/src_fixes_spiderTrap.o"
 .importobj "obj/src_fixes_tag_anywhere.o"
+.importobj "obj/src_initialization_cosmetic.o"
+.importobj "obj/src_initialization_file.o"
+.importobj "obj/src_initialization_init.o"
+.importobj "obj/src_initialization_item_rando.o"
+.importobj "obj/src_initialization_qol.o"
+.importobj "obj/src_item rando_actors.o"
+.importobj "obj/src_item rando_defs.o"
+.importobj "obj/src_item rando_flags.o"
+.importobj "obj/src_item rando_item_grab.o"
+.importobj "obj/src_item rando_spawning.o"
+.importobj "obj/src_item rando_text.o"
 .importobj "obj/src_misc_aesthetic.o"
 .importobj "obj/src_misc_beaver.o"
 .importobj "obj/src_misc_cutscene_database.o"
@@ -36,7 +46,6 @@
 .importobj "obj/src_misc_win_condition.o"
 .importobj "obj/src_randomizers_boss_rando.o"
 .importobj "obj/src_randomizers_every_lz_rando.o"
-.importobj "obj/src_randomizers_item_rando.o"
 .importobj "obj/src_randomizers_kut_out.o"
 .importobj "obj/src_randomizers_k_rool_order.o"
 .importobj "obj/src_randomizers_level_order.o"
